third_party/blink/web_tests/fast/dom/shadow/shadow-root-new.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p>This test checks the object 'host.attachShadow({mode: 'open'})' and internals.shadowRoot(div) are the same.</p>
<pre id="console"></pre>

<script>
function doTest() {
    if (!window.internals) {
        debug('This test runs on DRT only');
        return;
    }

    div = document.createElement('div');
    root1 = div.attachShadow({mode: 'open'});
    root2 = internals.shadowRoot(div);

    shouldBe('root1', 'root2');
}

doTest();
var successfullyParsed = true;
</script>
</body>
</html>
